Move `inviteUsersToRoom` to `RoomUpgrade`
This method is only used in one place, uses only public methods, and is
undocumented. Let's move it to the place where it is used, to simplify the API
for `RoomInvite`.
